retail-agent/requirements.txt
# Retail Agent Dependencies
# Python 3.10+

# HTTP and HTML parsing
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Supabase client
supabase>=2.0.0

# Async utilities  
asyncio-pool>=0.6.0
aiofiles>=23.0.0

# Environment
python-dotenv>=1.0.0

# Optional: monitoring
prometheus-client>=0.19.0

# LLM Agent (NVIDIA AI Endpoints)
langchain-nvidia-ai-endpoints>=1.0.0
langchain-core>=1.0.0

# Dashboard & Visualization
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
